% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neprior.R
\name{neprior}
\alias{neprior}
\title{Calculate numeric expectation for a given prior.}
\usage{
neprior(f = function(x) {     x }, g = function(x) {     1 },
  lower = -Inf, upper = Inf, ...)
}
\arguments{
\item{f}{A function, E[f(X)] will be returned. default: f(x) = x}

\item{g}{A function for kernel of prior. default: g(y) = 1}

\item{lower}{lower bound for y in g(y) function}

\item{upper}{upper bound for y in g(y) function}
}
\value{
E^\pi[f(X)], \pi: g(x).
}
\description{
Calculate numeric expectation for a given prior.
}
\examples{
neprior()

neprior(f = function(x) {x^2}, g = function(y) {y}, lower = -10, upper = 3 )

norm.kernel <- function(x) {exp(- (x - 2)^2 / (2 * 3^2))} # prior: N(2,3)
neprior(function(x) x^2, g = norm.kernel, lower = -Inf, upper = Inf) # E[X^2] = 9 + 2^2 = 13
}
